Part 11 Complete
Items can now be dropped on the ground under the player. Like the interaction point, it needs to be tweaked to be in the direction of the mouse. Perhaps once I figure out how to do the melee combat aiming at mouse I can apply the exact same logic to the interaction point and the dropping point.