45:39 Into Code Monkey Multiplayer Tutorial
Got through most of the multiplayer start tutorial. Players are able to connect to eachother but theres bugs such as the inventory completely shitting itself. I need to finish the tutorial and make it so that all of the inventory stuff is handled individually by each client (except the chests of course)